numeric extension move into core part 4
moved numeric/pixel_numeric_operations.hpp into core
