+++ test/core/pixel/Jamfile
@@ -1,6 +1,7 @@
 # Boost.GIL (Generic Image Library) - tests
 #
 # Copyright (c) 2019 Mateusz Loskot <mateusz@loskot.net>
+# Copyright (c) 2021 Pranam Lashkari <plashkari628@gmail.com>
 #
 # Distributed under the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE_1_0.txt or
@@ -19,3 +20,5 @@ compile pixels_are_compatible.cpp ;
 run color_convert_cmyk.cpp ;
 run packed_pixel.cpp ;
 run test_fixture.cpp ;
+run pixel_numeric_operations.cpp ;
+run pixel_numeric_operations_float.cpp ;
