+++ test/extension/numeric/Jamfile
@@ -17,7 +17,5 @@ compile-fail kernel_1d_fixed_even_size_fail.cpp ;
 run kernel.cpp ;
 run kernel_fixed.cpp ;
 run matrix3x2.cpp ;
-run pixel_numeric_operations.cpp ;
-run pixel_numeric_operations_float.cpp ;
 run resample.cpp ;
